docs(parts-calculator): move framing Notes above Cut plan
Requested by barthel.
