+++ parts-calculator/src/lib/components/CutPlan.svelte
@@ -4,14 +4,8 @@
 	import { layerStore } from '$lib/layers.svelte';
 	import { ftin, expand, packOptimal, packBundle, planGroups } from '$lib/cutplan';
 	import Popover from '$lib/components/Popover.svelte';
-	import Callout from '$lib/components/Callout.svelte';
-	import Figure from '$lib/components/Figure.svelte';
 	import { framingCsv } from '$lib/parts-csv';
 	import { download, exportSpec, filename } from '$lib/csv';
-
-	// lengths quoted in the C/D note below, read off the pieces themselves
-	const lenC = FRAMING_PIECES.find((p) => p.letter === 'C')?.len ?? 0;
-	const lenD = FRAMING_PIECES.find((p) => p.letter === 'D')?.len ?? 0;
 
 	let stock = $state(STOCK_MM);
 	let kerf = $state(3);
@@ -323,41 +317,6 @@
 			{/each}
 		</div>
 	</div>
-
-	<!-- notes: sat at the very bottom of the page, under the cut sheet, where
-	     nobody read them. They matter before you start cutting, so they go here -->
-	<Callout variant="info" title="Notes">
-		<ul class="list-disc space-y-2 pl-4 leading-relaxed">
-			<li>
-				<b class="text-text">Not all 2020 extrusion is the same stock.</b> Builders have found that
-				secondhand 20×20 mm profile varies: one with more angular, square corners on the T-slot seats
-				tightly against the crossbeam brackets, while a profile with more rounded corners can fit
-				loosely and add friction. Check the corner profile before buying, especially secondhand.
-				<div class="mt-2 max-w-xs">
-					<Figure
-						src="https://assets.basically.website/sorter-parts/img-0036-full-76e2483ec1e7.jpg"
-						alt="Two 2020 extrusion profiles held end-on side by side — the right one has angular T-slot corners, the left one has rounded corners"
-						caption="Right: angular corners, seats tight. Left: rounded corners, fits loose."
-						imgClass="max-h-40"
-					/>
-				</div>
-			</li>
-			<li>
-				<b class="text-text">D stands in for C at the bottom.</b> Every layer above the bottom two
-				gets 6 layer supports (<b class="text-text">C</b>, {lenC} mm). The bottom two share 6 foot
-				extensions (<b class="text-text">D</b>, {lenD} mm), one spanning both, in place of a C on
-				each. So a 1 or 2 layer build has no C in the list at all, and that is not a missing piece.
-			</li>
-			<li>
-				Pieces that share a cut length stack together at the saw — mark and cut the top bar, the rest
-				follow: <b class="text-text">A &amp; G</b> = 320 mm, <b class="text-text">B &amp; H</b> = 158 mm.
-			</li>
-			<li>
-				Where the cut length is under the CAD length, the piece is trimmed {CLEARANCE_MM} mm for
-				tolerance (see the <b class="text-text">Cut length</b> note above).
-			</li>
-		</ul>
-	</Callout>
 
 	<!-- cut sheet -->
 	<div>
